#2131 - hook up getAllTeamTypes
