+++ src/backend/src/controllers/design-reviews.controllers.ts
@@ -13,6 +13,15 @@ export default class DesignReviewsController {
     }
   }
 
+  static async getAllTeamTypes(req: Request, res: Response, next: NextFunction) {
+    try {
+      const teamTypes = await DesignReviewsService.getAllTeamTypes();
+      return res.status(200).json(teamTypes);
+    } catch (error: unknown) {
+      next(error);
+    }
+  }
+
   static async deleteDesignReview(req: Request, res: Response, next: NextFunction) {
     try {
       const drId: string = req.params.designReviewId;
